+++ CHANGELOG.md
@@ -24,6 +24,8 @@
   - `TestEmpty()` now supports tracing
   - `TestFilepathExists()` now supports tracing
   - `TestNotEmpty()` now supports tracing
+* Pipelines now set a context flag in their Pipe
+  - this is used to tell sinks where to read from!
 
 ### Dependencies